Add test case test_digest_multiple() - verify if multiple updates is ok.
This commit addresses a question raised during the review, whether:
you don't need to load all files in memory to compute this hash -- you can load files one by one and update the hash

A new test-case, test_digest_multiple(), to verify whether the use of
digest_message() across each file will update the hash. The test case
shows that this is not the case, which is why we cannot apply the
suggested review comment.

Other minor cleanup ...